Constructor for ECGraph Objects

Creates an ECGraph object from a file, data frame, or igraph object.

Usage

ECGraph(edges_input, col1 = 1, col2 = 2, header = TRUE, sep = "\t")

Arguments

edges_input

A character string specifying the path to a file, a data frame with two columns representing edges, or an igraph graph object. Vertex IDs will be coerced to character.

col1

The name or index of the first column if `edges_input` is a data frame or file. Defaults to 1.

col2

The name or index of the second column if `edges_input` is a data frame or file. Defaults to 2.

header

A logical value for reading from a file. Defaults to TRUE.

sep

The field separator character for reading from a file. Defaults to "\t".

Value

An object of class ECGraph.

See also

Examples

# --- Data ---
edge_df <- data.frame(
   vertex1 = c("A", "A", "B", "C", "E"),
   vertex2 = c("B", "C", "C", "D", "E"),
   stringsAsFactors = FALSE
)
temp_file <- tempfile(fileext = ".tsv")
write.table(edge_df, temp_file, sep = "\t", row.names = FALSE, quote = FALSE)

# --- ECGraph from a data frame ---
graph_from_df <- ECGraph(edge_df)

# --- ECGraph from a file path ---
graph_from_file <- ECGraph(temp_file)

unlink(temp_file)