Constructor for ECTermScoring Objects

Creates an ECTermScoring object from a data frame or file.

ECTermScoring(term_element_edges, col_term = 1, col_element = 2, ...)

Arguments

term_element_edges

A data frame or path to a file. Needs two columns: one for term IDs and one for element IDs, representing membership. If a file path, it's assumed to be tab-separated with a header by default.

col_term

The name or index of the column containing term IDs (in `term_element_edges`). Defaults to 1.

col_element

The name or index of the column containing element IDs (in `term_element_edges`). Defaults to 2.

...

Additional arguments passed to `utils::read.table` if `term_element_edges` is a file path (e.g., `sep`, `header`).

Value

An object of class ECTermScoring.

See also

Examples

# Create a sample term-element membership data frame
te_df <- data.frame(
   term = c("TermA", "TermA", "TermB", "TermC", "TermC"),
   element = c("Elem1", "Elem2", "Elem2", "Elem3", "Elem4")
)

# Create an ECTermScoring object
ects <- ECTermScoring(te_df)
print(ects@terms)
#> [1] "TermA" "TermB" "TermC"
print(ects@elements)
#> [1] "Elem1" "Elem2" "Elem3" "Elem4"