Vectorized calculation of fast between statistics

Calculates fast between statistics for a collection of set pairs using vectorized algorithm.

Usage

calculate_between_stats_fast_vectorized(object, pairs_dt, set_membership_dt)

# S4 method for class 'ECProb'
calculate_between_stats_fast_vectorized(object, pairs_dt, set_membership_dt)

Arguments

object

An ECProb object.

pairs_dt

A data table of pairs of element sets, with two columns "set1" and "set2" giving the set IDs.

set_membership_dt

A data table mapping set IDs to their elements. Must have "set_id" and "element" columns.

Value

A final `data.table` with the original pairs and columns for all calculated statistics (`set1`, `size1`, `set2`, `size2`, `observed_edges`, `lambda`, `p_value`, `log2_Anscombe_ratio`).

Details

This function is designed for efficiency when processing many term pairs at once. It uses a series of `data.table` joins and aggregations to avoid slow R-level loops. Note that this method uses the "fast" lambda approximation.

Methods (by class)

calculate_between_stats_fast_vectorized(ECProb): Method for ECProb objects

Examples

# TBD