Get Max Edge Count Between Sets (Fixed Degree Sequence)

Calculates the maximum possible number of edges between two disjoint sets of vertices, given their prescribed degrees (Fixed Degree Sequence model).

Usage

get_edge_count_between_max_fds(object, set1, set2)

# S4 method for class 'ECGraph'
get_edge_count_between_max_fds(object, set1, set2)

Arguments

object

An ECGraph object.

set1

A character vector of vertex names.

set2

A character vector of vertex names.

Value

The maximum possible number of edges, while preserving vertex degrees.

Methods (by class)

get_edge_count_between_max_fds(ECGraph): Method for ECGraph objects.

Examples

# Bipartite graph example
edge_df <- data.frame(p1=c("A","A","B"), p2=c("X","Y","Y"))
ecg <- ECGraph(edge_df)
get_edge_count_between_max_fds(ecg, c("A","B"), c("X","Y")) # Returns 3
#> [1] 3