Get Maximum Possible Edge Count Within a Set

Calculates the maximum possible number of edges in a simple graph formed by a given set of vertices (i.e., a complete graph). Only vertices present in the ECGraph object are considered.

Usage

get_edge_count_in_max(object, vertices)

# S4 method for class 'ECGraph'
get_edge_count_in_max(object, vertices)

Arguments

object

An ECGraph object.

vertices

A character vector of vertex names.

Value

The maximum possible number of edges.

Methods (by class)

get_edge_count_in_max(ECGraph): Method for ECGraph objects.

Examples

ecg <- ECGraph(data.frame(p1=c("A","B","C","D"), p2=c("B","C","D","A")))
# The function ignores vertices not in the graph, like "X"
# It calculates max edges for the valid set {A, B, C}, which is 3.
get_edge_count_in_max(ecg, c("A", "B", "C", "X"))
#> [1] 3