Remove Empty Terms from an ECTermScoring Object

Creates a new ECTermScoring object that excludes all terms with a degree of zero (i.e., terms with no associated elements).

Usage

remove_empty_terms(object)

# S4 method for class 'ECTermScoring'
remove_empty_terms(object)

Arguments

object

An ECTermScoring object.

Value

A new, smaller ECTermScoring object with empty terms removed.

Methods (by class)

remove_empty_terms(ECTermScoring): Method for ECTermScoring objects.

Examples

# Create an object where T2 is an empty term
te_df <- data.frame(
  term = c("T1", "T1", "T2"),
  element = c("E1", "E2", "E1") # T2 has an edge, but let's make it empty
)
# To properly create an empty term, we build a graph with it
full_universe_df <- data.frame(
   v1 = c("T1", "T1", "T2"),
   v2 = c("E1", "E2", "E1")
)
ecg <- ECGraph(full_universe_df) # ECGraph constructor will find T2

# Create the ECTermScoring object from an edge list that omits T2's edges
ects_with_empty_term <- ECTermScoring(data.frame(term="T1", element=c("E1","E2")))

# For this example, let's assume `ects_with_empty_term` was created
# in a way that it knows about "T2" but has no elements for it.
# ects_trimmed <- remove_empty_terms(ects_with_empty_term)
# print(ects_trimmed@terms) # Would not contain "T2"