Summarize Graph's Suitability for Fast Lambda Approximation

Provides statistics on the pairwise Bernoulli parameters (p_ij) to help assess if fast lambda approximation methods are suitable for the graph. This method also reports the proportion of vertices that are problematic for fast lambda approximation in the sense that their degree is greater than sqrt(2*M), M being the graph size.

Usage

summarize_suitability_fast(object)

# S4 method for class 'ECProb'
summarize_suitability_fast(object)

Arguments

object

An ECProb object.

Value

A list of summary statistics: 'p_ij_over_1' = proportion of vertex pairs such that p_ij > 1 and 'prop_problematic_vertices' = proportion of vertices with degree > sqrt(2*M)

Methods (by class)

summarize_suitability_fast(ECProb): Method for ECProb objects