Score Terms Against a Ranked List of Elements

Calculates a running enrichment score for terms based on a ranked list of elements.

Usage

terms_ecranks_statistics(object, element_ranks)

# S4 method for class 'ECTermScoring'
terms_ecranks_statistics(object, element_ranks)

Arguments

object

An ECTermScoring object.

element_ranks

A named list or vector of element ranks.

Value

A named list where each element is a `data.table` of statistics for a term.

Details

This method implements a fast, vectorized algorithm. For statistical rigor, the analysis is conditioned on the universe of elements provided in the `element_ranks` argument.

Methods (by class)

terms_ecranks_statistics(ECTermScoring): Method for ECTermScoring objects.

Examples

# Load sample data included with the package
data(sample_ects)

# Create a ranked list using a subset of the elements
set.seed(1)
elements_subset <- sample(sample_ects@elements, 2000)
ranked_list <- setNames(seq_along(elements_subset), elements_subset)

# Run the analysis
ranked_scores <- terms_ecranks_statistics(sample_ects, ranked_list)

# View the results for a specific term
terms <- get_neighbors(sample_ects@ecprob, elements_subset[1])
term_sizes <- lengths(sample_ects@ecprob@adj[terms])
term <- terms[which.max(term_sizes)]
print(term)
#> [1] "GO:0010524"
print(ranked_scores[[term]])
#>    element_id global_rank rank_in_term observed_ec max_ec term_size      lambda
#>        <char>       <num>        <int>       <int>  <num>     <int>       <num>
#> 1:      22953           1            1           1      1         4 0.002174859
#> 2:       5027         187            2           2      4         4 0.198999565
#> 3:        581        1114            3           3      4         4 1.152675076
#> 4:        444        1140            4           4      4         4 1.172683776
#>        p_value log2_Anscombe_ratio
#>          <num>               <num>
#> 1: 0.002170139           0.9330631
#> 2: 0.017357441           1.0244030
#> 3: 0.104522391           0.5717749
#> 4: 0.024563948           0.7495861