Trim High-Degree Vertices from an ECGraph

Removes the highest-degree vertices (hubs) from an ECGraph object until the condition for fast lambda approximation is met to the prescribed precision 'threshold' (no greater than 1).

Usage

trim_ecgraph(object, threshold = 1)

# S4 method for class 'ECGraph'
trim_ecgraph(object, threshold = 1)

Arguments

object

An ECGraph object.

threshold

A numeric threshold for the trimming condition. The default value of 1 corresponds to the theoretical condition `k_max^2 >= 2M`. Lowering this value will potentially result in better '_fast' approximations.

Value

A list containing two elements:

`trimmed_graph`: A new, smaller ECGraph object with without high-degree vertices.

`removed_vertices`: A character vector of the vertex names that were removed.

Details

This function provides a way to prepare a graph for analyses that use the `_fast` lambda calculation methods. It works by: 1. Identifying the current maximum degree (`k_max`) in the graph. 2. Checking if `k_max^2 >= threshold * 2M`, where `M` is the current total number of edges. 3. If the condition is met, all vertices with degree `k_max` are removed. 4. The graph size `M` is updated, and the process repeats.

Methods (by class)

trim_ecgraph(ECGraph): Method for ECGraph objects.

Examples

# Create a star graph where 'A' is a hub
edge_df <- data.frame(
  from = "A",
  to = c("B", "C", "D", "E", "F")
)
ecg <- ECGraph(edge_df)

# The default condition is not met: k_max=5, M=5.  5*5 / (2*5) > 1
trimmed_result <- trim_ecgraph(ecg)

# The central hub 'A' was removed
print(trimmed_result$removed_vertices) # "A"
#> [1] "A"

# The new graph contains only the isolated outer nodes
print(trimmed_result$trimmed_graph@adj)
#> $B
#> character(0)
#> 
#> $C
#> character(0)
#> 
#> $D
#> character(0)
#> 
#> $E
#> character(0)
#> 
#> $F
#> character(0)
#>