Vectorized Scoring of Terms Against Multiple Element Sets

A vectorized function that calculates enrichment statistics for all terms against a collection of input element sets.

Usage

terms_ecset_statistics_vectorized(object, input_sets, lambda_method = "fast")

# S4 method for class 'ECTermScoring'
terms_ecset_statistics_vectorized(object, input_sets, lambda_method = "fast")

Arguments

object

An ECTermScoring object.

input_sets

An object defining the input element sets. This can be either:

A `data.table` with two columns: `set_id` and `element`.

A named `list` where names are the set IDs and values are character vectors of elements.

lambda_method

The method for lambda calculation. Only "fast" is supported for this vectorized function.

Value

A named list of `data.table`s. Each name corresponds to an `input_set_id`, and each `data.table` contains the connectedness statistics for all connected terms (see terms_ecset_statistics()).

Details

This function is designed for efficiency when testing many input sets against the term database at once.

Methods (by class)

terms_ecset_statistics_vectorized(ECTermScoring): Method for ECTermScoring objects.

Examples

# Load sample data included with the package
library(data.table)
data(sample_ects)

# --- Example 1: Using a data.table as input ---
input_dt <- data.table(
  set_id = c("Set_A", "Set_A", "Set_B"),
  element = c(sample_ects@elements[1], sample_ects@elements[2], sample_ects@elements[3])
)
results_dt <- terms_ecset_statistics_vectorized(sample_ects, input_dt)
print(results_dt$Set_A)
#>       set1       set2 observed_edge_count term_degree sum_degrees_set
#>     <char>     <char>               <int>       <num>           <num>
#>  1:  Set_A GO:0000002                   2           8              19
#>  2:  Set_A GO:0006206                   1           1              19
#>  3:  Set_A GO:0006213                   1           2              19
#>  4:  Set_A GO:0046074                   1           2              19
#>  5:  Set_A GO:0051969                   1           3              19
#>  6:  Set_A GO:1905333                   1           3              19
#>  7:  Set_A GO:0004645                   1           2              19
#>  8:  Set_A GO:0009032                   1           2              19
#>  9:  Set_A GO:0015853                   1           5              19
#> 10:  Set_A GO:0015866                   1           4              19
#> 11:  Set_A GO:0046902                   1           9              19
#> 12:  Set_A GO:0140021                   1           4              19
#> 13:  Set_A GO:1990544                   1           6              19
#> 14:  Set_A GO:1990845                   1           9              19
#> 15:  Set_A GO:0005757                   1           6              19
#> 16:  Set_A GO:0005471                   1           5              19
#> 17:  Set_A GO:0015207                   1           4              19
#> 18:  Set_A GO:0017077                   1           3              19
#>     max_possible_edges       lambda      p_value log2_Anscombe_ratio
#>                  <int>        <num>        <num>               <num>
#>  1:                  2 0.0022790656 2.591158e-06           1.3271118
#>  2:                  2 0.0002848832 2.848426e-04           0.9366868
#>  3:                  2 0.0005697664 5.696041e-04           0.9361394
#>  4:                  2 0.0005697664 5.696041e-04           0.9361394
#>  5:                  2 0.0008546496 8.542844e-04           0.9355924
#>  6:                  2 0.0008546496 8.542844e-04           0.9355924
#>  7:                  2 0.0005697664 5.696041e-04           0.9361394
#>  8:                  2 0.0005697664 5.696041e-04           0.9361394
#>  9:                  2 0.0014244160 1.423402e-03           0.9344998
#> 10:                  2 0.0011395328 1.138884e-03           0.9350459
#> 11:                  2 0.0025639488 2.560662e-03           0.9323193
#> 12:                  2 0.0011395328 1.138884e-03           0.9350459
#> 13:                  2 0.0017092992 1.707838e-03           0.9339540
#> 14:                  2 0.0025639488 2.560662e-03           0.9323193
#> 15:                  2 0.0017092992 1.707838e-03           0.9339540
#> 16:                  2 0.0014244160 1.423402e-03           0.9344998
#> 17:                  2 0.0011395328 1.138884e-03           0.9350459
#> 18:                  2 0.0008546496 8.542844e-04           0.9355924

# --- Example 2: Using a named list as input ---
input_list <- list(
  Set_A = c(sample_ects@elements[1], sample_ects@elements[2]),
  Set_B = c(sample_ects@elements[3])
)
results_list <- terms_ecset_statistics_vectorized(sample_ects, input_list)
print(results_list$Set_A)
#>       set1       set2 observed_edge_count term_degree sum_degrees_set
#>     <fctr>     <char>               <int>       <num>           <num>
#>  1:  Set_A GO:0000002                   2           8              19
#>  2:  Set_A GO:0006206                   1           1              19
#>  3:  Set_A GO:0006213                   1           2              19
#>  4:  Set_A GO:0046074                   1           2              19
#>  5:  Set_A GO:0051969                   1           3              19
#>  6:  Set_A GO:1905333                   1           3              19
#>  7:  Set_A GO:0004645                   1           2              19
#>  8:  Set_A GO:0009032                   1           2              19
#>  9:  Set_A GO:0015853                   1           5              19
#> 10:  Set_A GO:0015866                   1           4              19
#> 11:  Set_A GO:0046902                   1           9              19
#> 12:  Set_A GO:0140021                   1           4              19
#> 13:  Set_A GO:1990544                   1           6              19
#> 14:  Set_A GO:1990845                   1           9              19
#> 15:  Set_A GO:0005757                   1           6              19
#> 16:  Set_A GO:0005471                   1           5              19
#> 17:  Set_A GO:0015207                   1           4              19
#> 18:  Set_A GO:0017077                   1           3              19
#>     max_possible_edges       lambda      p_value log2_Anscombe_ratio
#>                  <int>        <num>        <num>               <num>
#>  1:                  2 0.0022790656 2.591158e-06           1.3271118
#>  2:                  2 0.0002848832 2.848426e-04           0.9366868
#>  3:                  2 0.0005697664 5.696041e-04           0.9361394
#>  4:                  2 0.0005697664 5.696041e-04           0.9361394
#>  5:                  2 0.0008546496 8.542844e-04           0.9355924
#>  6:                  2 0.0008546496 8.542844e-04           0.9355924
#>  7:                  2 0.0005697664 5.696041e-04           0.9361394
#>  8:                  2 0.0005697664 5.696041e-04           0.9361394
#>  9:                  2 0.0014244160 1.423402e-03           0.9344998
#> 10:                  2 0.0011395328 1.138884e-03           0.9350459
#> 11:                  2 0.0025639488 2.560662e-03           0.9323193
#> 12:                  2 0.0011395328 1.138884e-03           0.9350459
#> 13:                  2 0.0017092992 1.707838e-03           0.9339540
#> 14:                  2 0.0025639488 2.560662e-03           0.9323193
#> 15:                  2 0.0017092992 1.707838e-03           0.9339540
#> 16:                  2 0.0014244160 1.423402e-03           0.9344998
#> 17:                  2 0.0011395328 1.138884e-03           0.9350459
#> 18:                  2 0.0008546496 8.542844e-04           0.9355924